feat: add helper and rewrite copy function
