+++ src/fs/copy.js
@@ -1,28 +1,14 @@
 import { promises as fsPromises } from 'fs';
-import path from 'path';
+import { join } from 'path';
 
-export const doesPathExist = async (path) => {
-  try {
-    await fsPromises.access(path);
-
-    return true;
-  } catch (err) {
-    if ((err.code = 'ENOENT')) {
-      return false;
-    } else {
-      throw err;
-    }
-  }
-};
-
+import { doesPathExist, OperationError, getPath } from '../helpers/index.js';
 
 const copyFiles = async (sourcePath, destinationDir) => {
   try {
     const files = await fsPromises.readdir(sourcePath);
-
     files.forEach(async (file) => {
-      const sourceFilePath = path.join(sourcePath, file);
-      const destinationFilePath = path.join(destinationDir, file);
+      const sourceFilePath = join(sourcePath, file);
+      const destinationFilePath = join(destinationDir, file);
 
       await fsPromises.copyFile(sourceFilePath, destinationFilePath);
     });
@@ -32,14 +18,18 @@ const copyFiles = async (sourcePath, destinationDir) => {
 };
 
 const copy = async () => {
-  const copyFolderPath = './src/fs/files_copy';
-  const sourceFolderPath = './src/fs/files';
+  const sourceFolderName = 'files';
+  const copyFolderName = 'files_copy';
+
   try {
+    const sourceFolderPath = getPath(import.meta.url, sourceFolderName);
+    const copyFolderPath = getPath(import.meta.url, copyFolderName);
+
     const isSourcePathExist = await doesPathExist(sourceFolderPath);
     const isCopyPathExist = await doesPathExist(copyFolderPath);
 
     if (!isSourcePathExist || isCopyPathExist) {
-      throw new Error('FS operation failed');
+      OperationError();
     }
 
     await fsPromises.mkdir(copyFolderPath, { recursive: true });
